____              ____   ___ ____  _____ |  _ \  ___  ___  |___ \ / _ \___ \|___ / | | | |/ _ \/ __|   __) | | | |__) | |_ \ | |_| |  __/ (__   / __/| |_| / __/ ___) | |____/ \___|\___| |_____|\___/_____|____/
New shell variable: RestoreCwdAfterExec. When true (false by default), the cwd
can never be changed by an external command as the cwd is stored before and
restored after command execution.

How the environment (argv,envp) is constructed for internal and external
commands has been modified:
* FIX: The length of the command line buffer is stored as an 'int' by Gulam.
Length is stored as 'uchar' in the Pexec command line. Previously it was
assigned without checking for truncation.
* FIX: The maxium length of the DTA section command line has been redefined as
124 bytes (as per tos_hyp). Previously it was 126.
* The value 'mw' (Mark Williams) for the shell variable 'env_style' has been
renamed 'argv'. Since it implemented what became known as the basic argv
protocol anyway, I beleive this is clearer.
* The argv protocol is enabled by default.
* FIX: The argv protocol is now only used when enabled and necessary. This
improves compatibility with programs that don't understand the protocol.
Previously, if the argv env_style was enabled, the command line length embedded
in Pexec calls was always set to 127. This was always invalid for such programs.
* If the argv protocol is disabled but the command line is too long, the user
is prompted whether to continue.
* FIX. Previously the standard command lines (i.e. not argv) of external
commands were stripped of quotes meaning that quoted strings were tokenized on
spaces.

Changes to file type determination:
* APP and GTP are now considered executable.
* FIX: Previously, a directory that happened to have an executable extension
would have been identified as a file not a directory.

General changes:
* The cursor is made visible at the start of every command prompt.